Create Vehicle Detection System
The Proposed design is cost-effective, reliable and
has the function of accurate tracking .When accident occurs
there is no one to help at the spot of accident so for quick
response we introduced gps and gsm based security. Vibration
sensor is used for identify the severity of the accident. When
large object or vehicles were spread out over ground, the owner
corporations often found it difficult to keep track of what was
happening. They required some type of system to determine
where each object was at any given time and for how long it
travelled. Also the need of tracking in consumer’s vehicle use to
prevent any kind of theft because police can use tracking reports
to locate stolen vehicle. GSM and GPS based tracking system will
provide effective, real time vehicle location, and reporting. A
GPS-GSM based tracking system will inform where your vehicle
is and where it has been, how long it has been. The system uses
geographic position and time information from the Global
Positioning Satellites. The system has an On-Board Module & 
which resides in the vehicle to be tracked.